iuii

+++ facturable/models/models.py
@@ -5,6 +5,6 @@ from odoo import models, fields, api
 
 class facturable(models.Model):
     _inherit = 'sale.order.line'
-    facturable=fields.Double('Facturable')
-    
+    facturable=fields.Float('Facturable')
+
